Refactor/110 httponly cookie login api test (#111)
* 프론트에서 API 통과를 위한 코드 수정

* 프론트에서 API 통과를 위한 코드 수정
