+++ src/main/java/com/keunsori/keunsoriserver/domain/member/dto/response/MyPageResponse.java
@@ -5,13 +5,15 @@ import com.keunsori.keunsoriserver.domain.member.domain.Member;
 public record MyPageResponse(
         String name,
         String studentId,
-        String email
+        String email,
+        String status
 ) {
     public static MyPageResponse from(Member member){
         return new MyPageResponse(
                 member.getName(),
                 member.getStudentId(),
-                member.getEmail()
+                member.getEmail(),
+                member.getStatus().toString()
         );
     }
 }
